Refactoring of signIn
