+++ src/components/SignIn/styles.scss
@@ -1,0 +1,3 @@
+.socialSingin {
+  margin: 0.5rem auto 0;
+}
