--- src/components/SingIn/styles.scss
@@ -1,24 +1,0 @@
-.singin {
-  display: block;
-  width: 100%;
-  max-width: 40rem;
-  margin: 4rem auto 6rem;
-  border: 1px solid black;
-  .wrap {
-    padding: 10px;
-    h2 {
-      font-size: 2.2rem;
-      line-height: 1;
-      font-weight: 400;
-      text-transform: uppercase;
-      text-align: center;
-      display: block;
-      width: 100%;
-      padding: 0;
-      margin: 0 auto;
-    }
-    .socialSingin {
-      margin: 0.5rem auto 0;
-    }
-  }
-}
