+++ src/pages/Login/index.js
@@ -1,11 +1,11 @@
 import React from "react";
 import "./styles.scss";
-import SingIn from "../../components/SingIn";
+import SignIn from "../../components/SignIn";
 
 function Login(props) {
   return (
     <div>
-      <SingIn />
+      <SignIn />
     </div>
   );
 }
